Remove pull request template.
People are not using the pull request template properly. Often, the fields
remain unfilled, and the placeholders are not deleted. This is cognitively
taxing for reviewers. Also, most of our changes are better described in the
commit log than in the pull request description. And, our reviews are quite
cursory and don't need a detailed pull request description for the reviewer to
act on.

* .github/PULL_REQUEST_TEMPLATE.md: Delete file.