Use multiprocessing to speed up computation
This commit refactors the code to make it possible to use multiprocessing to
speed up the computation of the partial correlations.

The major refactor is to move the `__compute_trait_info__` function to the
top-level of the module, and provide to it all the other necessary context via
the new args.