db: traits: Allow insertions of "0" in resp. sample values tables
When editing values from "x" to "0"(or any other value) when editing data, an
"update" statement was being run; thereby no new value was being inserted.  To
the end user, modifying an "x" value to something else meant that no value was
being inserted.  This commit fixes that by doing an insert whenever a change
from "x" to "0" is performed.

* gn3/db/traits.py (update_sample_data): Add insert statements whenever an
"update" statement returns a 0 row-count.