db: traits: Enable deletion of individual sample values
If a user replaces an individual value with an "x", delete that date entry
from the respective table.  Deletion here is the only option since by default
the Nstrain, PublishData and PublishSE don't accept null values.  Note that
deleting all 3 values is equivalent to removing the sample from the CSV file.

* gn3/db/traits.py (update_sample_data): If a value is "x", delete it from the
respective table.