Fix issue where 0's were treated as False for the primary trait in correlations
In the original version of the if statement* I believe it was
interpreted as "if a_val and (b_val is not None)". This caused
values of 0 for a_val (the primary trait's values) to be evaluated as
False.

I changed it to compare both a_val and b_val to None. This seems to have
fixed the issue.

* if (a_val and b_val is not None)