Add endpoint for checking state of external processes
Long-running computations are handed off to external processes. This avoids
timeouts in the webserver, and also reduces chances of instability of the
webserver.

The results of these long-running computations are needed eventually, so this
commit provides a way to check for the state of the computation, and the
results if any.