Make creation of database connections more flexible
* Pass the URI string to parse into `gn3.db_utils.parse_db_url` rather than
  relying on a global variable.
* Pass the URI string to use to generate the database connection to the
  `gn3.db_utils.database_connector` function rather than depending on a global
  variable. Use the global `SQL_URI` variable as a default value if one is not
  provided when calling the function.

The changes above make the creation of the database connection more flexible,
since the database URI passed into the function can be changed at the site
where the call is made.

The port is also parsed, and used where present, to allow for either a socket
connection or one based on a port.