Enable using different server for test database
This commit enables the tests to be run in separate test database located on a
completely separate database server from the live database, should that be
needed.

The motivation for this was because the database server on Penguin2 was slow
on cleanup of the test databases, making test development dreary.