Revert "Run json.loads on request.get_json, since request.get_json was just returning a string"
This reverts commit b93b22386056347d8002dd2e403425beeb4657cd.

The appropriate fix should have been in GN2. The original statement

    args = request.get_json()

was correct, since `request.get_json()` should return a python object parsed
from the JSON string in the request. Unfortunately, GN2 was encoding the
request data two times, which led to the call returning a JSON-encoded string
instead of the expected object.

The issue has been fixed in GN2 and therefore, the "fix" here can be reverted.