auth: Do not treat not finding a group as an error
Stop treating not finding a group for a user as an error, since that is an
actual expected state in the system at some point. Rather, restore the use of
Maybe rather than using Either.

We can raise the NotFoundError at the API level rather than at database
retrival level.