Enable use of `database_connection` in scripts without current_app
There is need to run external scripts using the same configurations as the
application but without the need to couple the script to the application.

In this case, we provide the needed configuration directly in the CLI, and
modify the existing `gn3.db_utils.database_connection` function to allow it to
work coupled to the app or otherwise.