auth: Fix bug with migration of user collections
The code, as written previously had a subtle bug - if the user created a new
collection before they had tried accessing their list of collections, the
older code would not have migrated the older collections.

This commit fixes that by enabling the migration of older collections, whether
or not the user has created a collection with their new accounts.