auth: Change check for client secret
We are saving the client secret in an encrypted form, meaning we have to
verify that the CLIENT_SECRET that is provided is the same one as was
generated at registration in a different way.

Initially, I was doing a direct comparison, having saved the CLIENT_SECRET
value as unencrypted plain-text.