Bug: Set the `data_exists` variable within the context manager
The `cursor.fetchone()` call that was used as the condition to the `if` was
called outside of the context manager, and therefore would always give a
non-truthy value at best and an inconsistent result at worst.

This commit gets the value before the context manager has exited and stores it
for later use.