Bug: Fetch group name by dataset_id
Change the code to fetch the group name by the dataset ID, since according to
usage[1] of the `retrieve_group_name` function, the value passed in is the
`dataset_id` and not the `group_id`.

Change the name from `retrieve_group_name` to `retrieve_phenotype_group_name`
to more clearly indicate that this function concerns itself with the groups
that relate to phenotypes.

[1](https://github.com/genenetwork/genenetwork2/blob/1bbb0430732b7fa5102d7dcbda80ebda252f5424/wqflask/wqflask/metadata_edits.py)