Check for a running process by viewing the build dir's contents.
In the CI build, the actual build is run in the
xapian_directory/build, which is seen as the xapian_directory in this
script.  The CI handles clean up WRT removing files related to the
build process.

* scripts/index-genenetwork (create_xapian_index): Create the xapian
directory if it doesn't exist.  If the xapian directory has files,
exit.  Create the temporary directory inside the xapian_directory.
Remove "build_directory.rmdir()"

Signed-off-by: Munyoki Kilyungi <me@bonfacemunyoki.com>