Bug: Use absolute path in place of relative path
Prior to this commit, an assumption is made that the application is
invoked with the root of the repository being the current working
directory. If this assumption is not met, then the system will not
find the appropriate path.

This commit forces the use of an absolute path, which should help
avoid issues with misplaced relative paths.