APPLICATION_ENVIRONMENT: New configuration variable
Add a new application configuration variable,
"APPLICATION_ENVIRONMENT" that will be useful to distinguish resources
for different instances of the application that might run on the same
host. The default value is an empty string in order to allow backwards
compatibility with existing code.

An example of where this finds utility: If you run multiple instances
of the application, but they all share the same redis instance, then
key-collisions from the different running applications will happen,
leading to subtle, difficult to troubleshoot bugs.